Fix bugs, make things work for open slots

+++ Heze/src/main/java/edu/wpi/cs3733/heze/entity/ScheduleTime.java
@@ -4,7 +4,7 @@ public class ScheduleTime {
 	long hour;
 	long minute;
 	
-	public ScheduleTime(int hour, int minute) {
+	public ScheduleTime(long hour, long minute) {
 		this.hour = hour;
 		this.minute = minute;
 	}
